refactor: form 동적 생성 로직 분리
